requirements.txt
tensorflow>=1.1.0
flask>=0.12.2
pillow>=4.1.1